Initial version of the full game

+++ scripts/classes.js
@@ -38,6 +38,8 @@ function Entity(texture, colour, power, speed, team, x, y) {
     
     this.tint = this.colour;
     
+    this.power = power;
+    
     this.interactive = true;
     this.hitArea = new PIXI.Rectangle(0, 0, 20, 20);
     
@@ -51,7 +53,8 @@ function Entity(texture, colour, power, speed, team, x, y) {
     this.weaponTarget = new PIXI.Point();
     
     this.team = team;
-    this.weapon = new Weapon(this);
+    this.weapon = new Weapon(this, this.power);
+    this.armour = new Armour(this, this.power);
     
     if (this.team == 0) {
         PlayerAI.call(this);
@@ -71,12 +74,14 @@ function Entity(texture, colour, power, speed, team, x, y) {
     },this);
 }
 
-function Weapon(entity) {
+function Weapon(entity, power) {
     //A weapon is what creates bullets.
     
-    this.type = weaponTypes[0];
+    this.type = weaponTypes[Math.floor(Math.random() * weaponTypes.length)];
     this.bulletTexture = app.renderer.generateTexture(this.type.image);
     this.entity = entity;
+    
+    this.damage = power * this.type.damageMod;
     
     this.maxUse = this.type.useTime;
     this.curUse = this.maxUse;
@@ -84,7 +89,6 @@ function Weapon(entity) {
         if (this.curUse === 0) {
             app.bullets.push(new Bullet(this.entity, this.bulletTexture, 
             function() {
-                this.accel = setAccelInDirection(this.direction, this.speed);
                 this.tint = entity.colour;
             }, []));
             
@@ -101,9 +105,38 @@ function Weapon(entity) {
     app.ticker.add(this.reload, this);
 }
 
+function Armour(entity, power) {
+    this.entity = entity;
+    
+    this.maxHP = power * 10;
+    this.curHP = this.maxHP;
+    
+    this.maxRegen = 1.2;
+    this.curRegen = 1;
+    
+    app.ticker.add(function() {
+        if (app.tick % 6 == 0) {
+            if (this.curHP < this.maxHP) {
+                this.curHP *= this.curRegen;
+                if (this.curHP > this.maxHP) {
+                    this.curHP = this.maxHP;
+                }
+            } 
+            if (this.curRegen < this.maxRegen) {
+                this.curRegen *= 1.0001;
+                if (this.curRegen > this.maxRegen) {
+                    this.curRegen = this.maxRegen;
+                }
+            }
+        }
+    }, this);
+}
+
 function Bullet(entity, texture, moveFunction, moveConsts) {
     PIXI.Sprite.call(this, texture);
     this.entity = entity;
+    
+    this.damage = this.entity.weapon.damage;
     
     Moveable.call(this, this.entity.weapon.type.speed);
     this.move = moveFunction;
@@ -117,6 +150,10 @@ function Bullet(entity, texture, moveFunction, moveConsts) {
     this.direction = getAngleInRadians(this.position, this.entity.weaponTarget);
     
     this.position.copy(moveToPoint(entity.position, this.entity.weaponTarget, 10));
+    
+    this.accel = setAccelInDirection(this.direction, this.speed);
+    this.accel.x += this.entity.accel.x;
+    this.accel.y += this.entity.accel.y;
     
     this.tick = function(bullets, i) {
         this.curLifetime -= 1;
@@ -134,20 +171,36 @@ function Bullet(entity, texture, moveFunction, moveConsts) {
         if(this.entity.team == 0) {
             for (var n = 0; n < app.enemies.length; n += 1) {
                 if (app.enemies[n].hitArea.contains(this.position.x, this.position.y)) {
-                    app.enemies[n].delete();
-                    app.enemies.splice(n, 1);
-                    n -= 1;
+                    app.enemies[n].armour.curHP -= this.damage;
+                    if (app.enemies[n].armour.curHP <= 0) {
+                        app.money.moneyGainedIn5Sec[app.money.moneyGainedSec] += app.enemies[n].power;
+                        app.wave.enemiesOnScreen -= 1;
+                        app.enemies[n].delete();
+                        app.enemies.splice(n, 1);
+                        n -= 1;
+                    }
                     bullets.splice(i, 1);
                     app.stage.removeChild(this);
                 }
             }
-        }
-        
-        if(this.entity.team == 1) {
+        } else {
             if (app.player.hitArea.contains(this.position.x, this.position.y)) {
-                bullets.splice(i, 1);
-                app.stage.removeChild(this);
-            }
+                    app.player.armour.curHP -= this.damage;
+                    if (app.player.armour.curHP <= 0) {
+                        app.wave.number = 0;
+                        app.wave.enemiesInWave = 1;
+                        app.wave.enemiesOnScreen = 0;
+                        for (var n = 0; n < app.enemies.length; n += 1) {
+                            app.enemies[n].delete();
+                            app.enemies.splice(n, 1);
+                            n -= 1;
+                        }
+                        app.power = 1;
+                        app.player.armour.curHP = app.player.armour.maxHP;
+                    }
+                    bullets.splice(i, 1);
+                    app.stage.removeChild(this);
+                }
         }
     }
